+++ glimpse/core/src/desktopMain/kotlin/types/Colors.kt
@@ -23,8 +23,14 @@ private const val VEC4_SIZE = 4
 
 /**
  * Returns a [Vec3] representation of a given [color].
+ *
+ * @deprecated Use [toVec3] instead.
  */
 @Suppress("FunctionNaming")
+@Deprecated(
+    message = "Use Color.toVec3() instead",
+    replaceWith = ReplaceWith("color.toVec3()", "graphics.glimpse.types.toVec3")
+)
 fun Vec3(color: Color): Vec3<Float> {
     val output = FloatArray(size = VEC3_SIZE)
     color.getRGBColorComponents(output)
@@ -33,10 +39,54 @@ fun Vec3(color: Color): Vec3<Float> {
 
 /**
  * Returns a [Vec4] representation of a given [color].
+ *
+ * @deprecated Use [toVec4] instead.
  */
 @Suppress("FunctionNaming")
+@Deprecated(
+    message = "Use Color.toVec4() instead",
+    replaceWith = ReplaceWith("color.toVec4()", "graphics.glimpse.types.toVec4")
+)
 fun Vec4(color: Color): Vec4<Float> {
     val output = FloatArray(size = VEC4_SIZE)
     color.getRGBComponents(output)
     return Vec4.fromList(output.toList())
 }
+
+/**
+ * Returns a [Vec3] representation of this [Color].
+ *
+ * @since v1.3.0
+ */
+fun Color.toVec3(): Vec3<Float> {
+    val output = FloatArray(size = VEC3_SIZE)
+    getRGBColorComponents(output)
+    return Vec3.fromList(output.toList())
+}
+
+/**
+ * Returns a [Vec4] representation of this [Color].
+ *
+ * @since v1.3.0
+ */
+fun Color.toVec4(): Vec4<Float> {
+    val output = FloatArray(size = VEC4_SIZE)
+    getRGBComponents(output)
+    return Vec4.fromList(output.toList())
+}
+
+/**
+ * Returns a Compose [Color] representation of this [Vec3].
+ *
+ * @since v1.3.0
+ */
+fun Vec3<Float>.toColor(): Color =
+    Color(this.r, this.g, this.b)
+
+/**
+ * Returns a Compose [Color] representation of this [Vec4].
+ *
+ * @since v1.3.0
+ */
+fun Vec4<Float>.toColor(): Color =
+    Color(this.r, this.g, this.b, this.a)